refactor: rename src/auth.rs to src/anthropic_oauth.rs
The existing `auth` module held the Anthropic OAuth PKCE flow. Phase 1
of the Entra ID authentication rollout needs the `auth` namespace for a
new module-tree covering JWT validation, AuthContext, middleware, and
config. Renaming the existing flat module frees that namespace while
keeping its single responsibility (Anthropic OAuth token lifecycle)
clearly scoped in the new name.

No behavioral change. 10 call sites updated in src/config.rs,
src/config/load.rs, src/config/onboarding.rs, src/llm/manager.rs.

`pub mod auth;` is temporarily removed from lib.rs and restored in
Task 1.7 after the Entra submodules exist.

Precursor to Phase 1 Task 1.3 per
.scratchpad/plans/entraid-auth/phase-1-jwt-middleware.md.
